third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/require-sri-for/require-sri-for-script-allowed-meta.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta http-equiv="Content-Security-Policy" content="require-sri-for script; script-src 'self' 'unsafe-inline'">
<script>
    var executed_test = async_test("Script that requires integrity executes and does not generate a violation report.");
    document.addEventListener('securitypolicyviolation', executed_test.unreached_func("No report should be generated."));
</script>
<script crossorigin integrity="sha384-SOGIJ0vOWzweNE6RLF/TOXGmPzCxF5+dNuBP4x1NgnKsfC4yFCVIDJILalTMwUrp" src="ran.js"></script>
<script>
	assert_equals(z, 13);
    executed_test.done();
</script>
